requirements.txt
lxml>=3.3.5
PyYAML>=3.10
psycopg2>=2.5.1
ipython>=0.10
coverage==3.7.1
django-nose==1.2
nosexcover==1.0.10
fabric==1.8.3
geopy==0.99
markdown==2.4.1
python-mimeparse==0.1.4
mock==1.0.1
pylint==1.2.1
xlrd==0.9.3
pyPdf==1.13
reportlab==3.1.44
south==0.8.4
sphinx==1.2.2
unittest2==0.5.1
feedparser==5.1.3
gdata==2.0.18
manifesto==0.5
backports.ssl_match_hostname==3.4.0.2
django==1.6.7
django-debug-toolbar==1.2.1
django-tastypie==0.11.1
django-mptt==0.6.1
django-form-utils==1.0.1
django-pipeline==1.3.24
pilkit==1.1.12
redis==2.9.1
pysolr==3.2.0
elasticsearch==1.0.0
queues==0.6.3
queued_search==2.1.0
django-haystack==2.2.1
django-messages==0.5.0
httplib2==0.9
oauth2==1.5.211
django-storages==1.1.8
boto==2.29.1
django-fsm==2.1.0
pinder==1.0.1
django-qsstats-magic==0.7.2
-e git://github.com/e-loue/logbook.git#egg=logbook
-e git://github.com/e-loue/pyke.git#egg=pyke
-e git://github.com/e-loue/freshen.git#egg=freshen
-e git://github.com/e-loue/django-faq.git@haystack_2.x#egg=faq
-e git+https://github.com/ntucker/django-announcements.git#egg=announcements
-e git://github.com/e-loue/django-lean.git#egg=django_lean
-e git://github.com/e-loue/django-lint.git#egg=django-lint
-e git://github.com/e-loue/python-paypalx.git#egg=paypalx
-e git://github.com/e-loue/proclaim.git#egg=proclaim
-e git://github.com/e-loue/rollout.git#egg=rollout
-e git://github.com/e-loue/python-sdk.git#egg=facebook-sdk
-e git://github.com/e-loue/django-imagekit.git#egg=imagekit
beautifulsoup4
google-api-python-client
django-markup-deprecated
simplejson
python-dateutil==1.5
djangorestframework==2.3.14
djangorestframework-gis==0.2
django-filter==0.7
django-oauth2-provider
requests>=2.9.1
pyjade==2.2.0
wsgiref==0.1.2
static==1.0.2
dj-static==0.0.6
dj-database-url==0.3.0
django-getenv==1.3.1
django-redis-cache==0.13.0
django-iban==0.3.0
django-sslify==0.2.3
suds-jurko==0.6
-e git://github.com/dgrtwo/ParsePy.git#egg=parsepy
django-modeltranslation
sendgrid
django-import-export==0.4.2
openpyxl==2.3.2
https://cdn.mysql.com/Downloads/Connector-Python/mysql-connector-python-2.0.4.tar.gz
algoliasearch==1.8.0
-e git+https://github.com/matagus/haystack-algolia.git#egg=haystack_algolia
django-angular==0.7.15
pytest==2.9.0
pytest-django==2.9.1
fake-factory==0.5.7
factory_boy==2.6.1
analytics-python==1.2.3
winpdb
